pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-context-local"
version = "0.1.0"
description = "Local semantic code search for Claude Code (MCP) using EmbeddingGemma, FAISS, and AST/tree-sitter chunking. 100% local embeddings and indexing."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Farhan Ali Raza" }
]
maintainers = [
    { name = "Farhan Ali Raza" }
]
keywords = [
    "semantic-search",
    "code-search",
    "embeddings",
    "faiss",
    "mcp",
    "claude",
    "tree-sitter",
    "chunking",
    "python",
]
dependencies = [
    "click>=8.2.1",
    "faiss-cpu>=1.12.0",
    "fastmcp>=2.14.1",
    "huggingface-hub>=0.34.4,<1.0",
    "pyyaml>=6.0.1",
    "mcp>=1.24.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "rich>=14.2.0",
    "sentence-transformers[onnx]>=5.1.0",
    "sqlitedict>=2.1.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-c>=0.24.1",
    "tree-sitter-c-sharp>=0.23.1",
    "tree-sitter-cpp>=0.23.4",
    "tree-sitter-go>=0.25.0",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-html>=0.23.2",
    "tree-sitter-css>=0.25.0",
    "tree-sitter-json>=0.24.8",
    "tree-sitter-language-pack>=0.13.0",
    "tree-sitter-markdown>=0.3.2",
    "tree-sitter-yaml>=0.7.2",
    "tree-sitter-toml>=0.6.0",
    "tree-sitter-xml>=0.7.0",
    "tree-sitter-graphql>=0.1.0",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-rust>=0.24.0",
    "tree-sitter-svelte>=1.0.2",
    "tree-sitter-typescript>=0.23.2",
]

classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
]

[project.urls]
Homepage = "https://github.com/FarhanAliRaza/claude-context-local"
Repository = "https://github.com/FarhanAliRaza/claude-context-local"
Documentation = "https://github.com/FarhanAliRaza/claude-context-local#readme"
Issues = "https://github.com/FarhanAliRaza/claude-context-local/issues"

[tool.setuptools.packages.find]
include = ["claude_embedding_search*"]
